Implement FloatingIsland and add edit button
