+++ src/components/Applause/styles.css
@@ -1,28 +1,3 @@
-.applause-container {
-  height: 40px;
-  width: 110px;
-  align-items: center;
-  display: flex;
-  box-shadow: inherit;
-  border-radius: 20px;
-  background-color: var(--ifm-color-emphasis-200);
-  box-shadow: rgb(0 0 0 / 10%) 0px 2px 10px 0px;
-  z-index: calc(var(--ifm-z-index-fixed) - 1);
-  position: fixed;
-  bottom: 1.3rem;
-  right: calc(100% / 2);
-  left: calc(100% / 2.1);
-  justify-content: center;
-  opacity: 1;
-}
-
-@media screen and (max-width: 997px) {
-  .applause-container {
-    right: calc(100% / 2);
-    left: calc(100% / 2.66);
-  }
-}
-
 .applause-button {
   border: none;
   flex-direction: row;
