+++ src/components/EditThisPageButton/styles.css
@@ -1,0 +1,3 @@
+.edit-button > svg {
+  fill: var(--ifm-font-color-base);
+}
