+++ src/components/FloatingIsland/styles.css
@@ -1,0 +1,19 @@
+.applause-container {
+  height: 40px;
+  width: 140px;
+  align-items: center;
+  display: flex;
+  box-shadow: inherit;
+  border-radius: 20px;
+  background-color: var(--ifm-color-emphasis-200);
+  box-shadow: rgb(0 0 0 / 10%) 0px 2px 10px 0px;
+  z-index: calc(var(--ifm-z-index-fixed) - 1);
+  position: fixed;
+  bottom: 1.3rem;
+  right: calc(100% / 2);
+  left: calc(100% / 2.1);
+  justify-content: space-between;
+  padding-left: 1.2rem;
+  padding-right: 1.2rem;
+  opacity: 1;
+}
